+++ addons/mail/controllers/__init__.py
@@ -1,3 +1,4 @@
 # -*- coding: utf-8 -*
 from . import main
 from . import bus
+from . import home
